Change POST to support future legacy fields
The new POST method is saved as before but the
object is wrapped using the
ProductModelAssembler.

The new return type ResponseEntity<*> is
determined at runtime (hence the
star-projection `*`). It will create an HTTP 201
Created status message instead of just returning
the saved product as before. The new return type
will include a Location response header, and
we use the URI derived from the model's
self-related link.

Also, it now returns the model equivalent of
Product instead of just the saved Product.