Change PUT to support future legacy fields
The new PUT method is saves as before and
replaces as before with the minor change that
if the ID of the new product is different from
the path variable ID then the new product will
be saved with the path variable ID instead of
the ID which the new product originally had.

Also, the new return type is ResponseEntity<*>
with a 201 Created (see previous commit for
further details) and it returns the model
equivalent of Product.